+++ src/timeline/stack.js
@@ -28,14 +28,14 @@ stack.orderByEnd = function orderByEnd(items) {
 };
 
 /**
- * Adjust vertical positions of the events such that they don't overlap each
+ * Adjust vertical positions of the items such that they don't overlap each
  * other.
  * @param {Item[]} items
  *            All visible items
  * @param {{item: number, axis: number}} margin
  *            Margins between items and between items and the axis.
  * @param {boolean} [force=false]
- *            If true, all items will be re-stacked. If false (default), only
+ *            If true, all items will be repositioned. If false (default), only
  *            items having a top===null will be re-stacked
  */
 stack.stack = function _stack (items, margin, force) {
@@ -68,11 +68,27 @@ stack.stack = function _stack (items, margin, force) {
         }
 
         if (collidingItem != null) {
-          // There is a collision. Reposition the event above the colliding element
+          // There is a collision. Reposition the items above the colliding element
           item.top = collidingItem.top + collidingItem.height + margin.item;
         }
       } while (collidingItem);
     }
+  }
+};
+
+/**
+ * Adjust vertical positions of the items without stacking them
+ * @param {Item[]} items
+ *            All visible items
+ * @param {{item: number, axis: number}} margin
+ *            Margins between items and between items and the axis.
+ */
+stack.nostack = function nostack (items, margin) {
+  var i, iMax;
+
+  // reset top position of all items
+  for (i = 0, iMax = items.length; i < iMax; i++) {
+    items[i].top = margin.axis;
   }
 };